requirements.txt
streamlit
python-docx
pymupdf